+++ miniconf/src/array.rs
@@ -44,9 +44,9 @@ fn get_mut<'a, const N: usize, K: Keys, T>(
 macro_rules! depth {
     ($($y:literal)+) => {$(
         impl<T: TreeKey<{$y - 1}>, const N: usize> TreeKey<$y> for [T; N] {
-            fn walk<W: Walk>() -> Result<W, W::Error> {
+            fn traverse_all<W: Walk>() -> Result<W, W::Error> {
                 W::internal().merge(
-                    &T::walk::<W>()?,
+                    &T::traverse_all::<W>()?,
                     None,
                     &KeyLookup {
                         len: N,
@@ -117,7 +117,7 @@ depth!(2 3 4 5 6 7 8 9 10 11 12 13 14 15 16);
 
 // Y == 1
 impl<T, const N: usize> TreeKey for [T; N] {
-    fn walk<W: Walk>() -> Result<W, W::Error> {
+    fn traverse_all<W: Walk>() -> Result<W, W::Error> {
         W::internal().merge(
             &W::leaf(),
             None,