+++ miniconf/src/option.rs
@@ -37,8 +37,8 @@ fn get_mut<'a, T, K: Keys>(
 macro_rules! depth {
     ($($y:literal)+) => {$(
         impl<T: TreeKey<{$y - 1}>> TreeKey<$y> for Option<T> {
-            fn walk<W: Walk>() -> Result<W, W::Error> {
-                T::walk()
+            fn traverse_all<W: Walk>() -> Result<W, W::Error> {
+                T::traverse_all()
             }
 
             fn traverse_by_key<K, F, E>(keys: K, func: F) -> Result<usize, Error<E>>
@@ -95,7 +95,7 @@ depth!(2 3 4 5 6 7 8 9 10 11 12 13 14 15 16);
 
 // Y == 1
 impl<T> TreeKey for Option<T> {
-    fn walk<W: Walk>() -> Result<W, W::Error> {
+    fn traverse_all<W: Walk>() -> Result<W, W::Error> {
         Ok(W::leaf())
     }