+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * `KeyLookup` has been changed from a trait to a struct.
-* `TreeKey::metadata() -> Metadata` has been changed to `walk<W: Walk>() -> W` with `impl Walk for Metadata`.
+* `TreeKey::metadata() -> Metadata` has been changed to `traverse_all<W: Walk>() -> W` with `impl Walk for Metadata`.
 
 ## [0.15.0](https://github.com/quartiq/miniconf/compare/v0.14.1...v0.15.0) - 2024-10-01